+++ space-d/src/main/java/com/dnd/spaced/core/account/domain/enums/JobGroup.java
@@ -1,6 +1,6 @@
-package com.dnd.spaced.core.account.domain;
+package com.dnd.spaced.core.account.domain.enums;
 
-import com.dnd.spaced.core.account.domain.exception.InvalidJobGroupException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidJobGroupException;
 import java.util.Arrays;
 import lombok.Getter;
