+++ space-d/src/main/java/com/dnd/spaced/core/account/domain/enums/exception/InvalidCompanyException.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.account.domain.exception;
+package com.dnd.spaced.core.account.domain.enums.exception;
 
 import com.dnd.spaced.global.exception.base.AccountClientException;
 import com.dnd.spaced.global.exception.code.AccountErrorCode;
