+++ space-d/src/main/java/com/dnd/spaced/core/account/domain/enums/exception/InvalidJobGroupException.java
@@ -1,11 +1,11 @@
-package com.dnd.spaced.core.account.domain.exception;
+package com.dnd.spaced.core.account.domain.enums.exception;
 
 import com.dnd.spaced.global.exception.base.AccountClientException;
 import com.dnd.spaced.global.exception.code.AccountErrorCode;
 
-public class InvalidExperienceException extends AccountClientException {
+public class InvalidJobGroupException extends AccountClientException {
 
-    public InvalidExperienceException(String message) {
+    public InvalidJobGroupException(String message) {
         super(AccountErrorCode.INVALID_EXPERIENCE, message);
     }
 }
