+++ space-d/src/test/java/com/dnd/spaced/config/docs/snippet/DocsController.java
@@ -4,10 +4,10 @@ import com.dnd.spaced.config.docs.snippet.dto.response.CommonDocsResponse;
 import com.dnd.spaced.config.docs.snippet.enums.EnumDocs;
 import com.dnd.spaced.config.docs.snippet.exceptions.ExceptionContent;
 import com.dnd.spaced.config.docs.snippet.exceptions.ExceptionDocs;
-import com.dnd.spaced.core.account.domain.Company;
-import com.dnd.spaced.core.account.domain.Experience;
-import com.dnd.spaced.core.account.domain.JobGroup;
-import com.dnd.spaced.core.account.domain.ProfileImageName;
+import com.dnd.spaced.core.account.domain.enums.Company;
+import com.dnd.spaced.core.account.domain.enums.Experience;
+import com.dnd.spaced.core.account.domain.enums.JobGroup;
+import com.dnd.spaced.core.account.domain.enums.ProfileImageName;
 import com.dnd.spaced.core.word.domain.Category;
 import com.dnd.spaced.core.word.domain.PronunciationType;
 import com.dnd.spaced.global.exception.code.AccountErrorCode;
