+++ space-d/src/test/java/com/dnd/spaced/core/account/application/AccountServiceTest.java
@@ -8,11 +8,11 @@ import com.dnd.spaced.config.clean.annotation.CleanUpDatabase;
 import com.dnd.spaced.core.account.application.dto.response.AccountInfoDto;
 import com.dnd.spaced.core.account.application.exception.ForbiddenAccountException;
 import com.dnd.spaced.core.account.domain.Account;
-import com.dnd.spaced.core.account.domain.ProfileImageName;
-import com.dnd.spaced.core.account.domain.exception.InvalidCompanyException;
-import com.dnd.spaced.core.account.domain.exception.InvalidExperienceException;
-import com.dnd.spaced.core.account.domain.exception.InvalidJobGroupException;
-import com.dnd.spaced.core.account.domain.exception.InvalidProfileImageNameException;
+import com.dnd.spaced.core.account.domain.enums.ProfileImageName;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidCompanyException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidExperienceException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidJobGroupException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidProfileImageNameException;
 import com.dnd.spaced.core.account.domain.repository.AccountRepository;
 import java.util.Arrays;
 import java.util.stream.Stream;
