+++ space-d/src/test/java/com/dnd/spaced/core/account/domain/AccountTest.java
@@ -5,13 +5,17 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.dnd.spaced.core.account.domain.exception.InvalidCompanyException;
+import com.dnd.spaced.core.account.domain.embed.CareerInfo;
+import com.dnd.spaced.core.account.domain.enums.Company;
+import com.dnd.spaced.core.account.domain.enums.Experience;
+import com.dnd.spaced.core.account.domain.enums.JobGroup;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidCompanyException;
 import com.dnd.spaced.core.account.domain.exception.InvalidIdException;
-import com.dnd.spaced.core.account.domain.exception.InvalidExperienceException;
-import com.dnd.spaced.core.account.domain.exception.InvalidJobGroupException;
-import com.dnd.spaced.core.account.domain.exception.InvalidNicknameException;
-import com.dnd.spaced.core.account.domain.exception.InvalidProfileImageException;
-import com.dnd.spaced.core.account.domain.exception.InvalidRoleNameException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidExperienceException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidJobGroupException;
+import com.dnd.spaced.core.account.domain.embed.exception.InvalidNicknameException;
+import com.dnd.spaced.core.account.domain.embed.exception.InvalidProfileImageException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidRoleNameException;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
