+++ space-d/src/test/java/com/dnd/spaced/core/account/domain/embed/ProfileInfoTest.java
@@ -1,12 +1,12 @@
-package com.dnd.spaced.core.account.domain;
+package com.dnd.spaced.core.account.domain.embed;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.dnd.spaced.core.account.domain.exception.InvalidNicknameException;
-import com.dnd.spaced.core.account.domain.exception.InvalidProfileImageException;
+import com.dnd.spaced.core.account.domain.embed.exception.InvalidNicknameException;
+import com.dnd.spaced.core.account.domain.embed.exception.InvalidProfileImageException;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
