+++ space-d/src/test/java/com/dnd/spaced/core/account/domain/enums/JobGroupTest.java
@@ -1,9 +1,9 @@
-package com.dnd.spaced.core.account.domain;
+package com.dnd.spaced.core.account.domain.enums;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.dnd.spaced.core.account.domain.exception.InvalidJobGroupException;
+import com.dnd.spaced.core.account.domain.enums.exception.InvalidJobGroupException;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
